:white_check_mark: refactor: actions 관련 공통 fixture를 conftest.py로 이동한다
